fixture refac
