+++ tests/conftest.py
@@ -1,9 +1,32 @@
 import copy
+import tempfile
 from pathlib import Path
-from typing import Callable
+from typing import Callable, Tuple
 
 import numpy as np
 import pytest
+import tifffile
+
+from careamics_restoration.config import Configuration
+from careamics_restoration.config.algorithm import Algorithm
+from careamics_restoration.config.data import Data
+from careamics_restoration.config.prediction import Prediction
+from careamics_restoration.config.training import LrScheduler, Optimizer, Training
+
+
+@pytest.fixture
+def image_size() -> Tuple[int, int]:
+    return (128, 128)
+
+
+@pytest.fixture
+def patch_size() -> Tuple[int, int]:
+    return (64, 64)
+
+
+@pytest.fixture
+def overlaps() -> Tuple[int, int]:
+    return (32, 32)
 
 
 @pytest.fixture
@@ -155,3 +178,56 @@ def array_3D() -> np.ndarray:
         3D array with shape (1, 5, 10, 9).
     """
     return np.arange(2048).reshape((1, 8, 16, 16))
+
+
+@pytest.fixture
+def temp_dir() -> Path:
+    with tempfile.TemporaryDirectory() as temp_dir:
+        yield Path(temp_dir)
+
+
+@pytest.fixture
+def example_data_path(
+    temp_dir: Path, image_size: Tuple[int, int], patch_size: Tuple[int, int]
+) -> Tuple[Path, Path]:
+    test_image = np.random.rand(*image_size)
+    test_image_predict = test_image[None, None, ...]
+
+    train_path = temp_dir / "train"
+    val_path = temp_dir / "val"
+    test_path = temp_dir / "test"
+    train_path.mkdir()
+    val_path.mkdir()
+    test_path.mkdir()
+
+    tifffile.imwrite(train_path / "train_image.tif", test_image)
+    tifffile.imwrite(val_path / "val_image.tif", test_image)
+    tifffile.imwrite(test_path / "test_image.tif", test_image_predict)
+
+    return train_path, val_path, test_path
+
+
+@pytest.fixture
+def base_configuration(temp_dir: Path, patch_size) -> Configuration:
+    configuration = Configuration(
+        experiment_name="smoke_test",
+        working_directory=temp_dir,
+        algorithm=Algorithm(loss="n2v", model="UNet", is_3D="False"),
+        data=Data(
+            data_format="tif",
+            axes="YX",
+        ),
+        training=Training(
+            num_epochs=1,
+            patch_size=patch_size,
+            batch_size=1,
+            optimizer=Optimizer(name="Adam"),
+            lr_scheduler=LrScheduler(name="ReduceLROnPlateau"),
+            extraction_strategy="random",
+            augmentation=True,
+            num_workers=0,
+            use_wandb=False,
+        ),
+        prediction=Prediction(use_tiling=False),
+    )
+    return configuration
